Add FoodNetwork Recipe Parser
Adds a parser interface which takes in an html node and returns a recipe
struct, but with the ingredients still as a string. This is in order to
scrape first and then clean the ingredients later.

It also adds a LinkSource interface which takes in an html page and
returns a list of links to other recipes to scrape, as well as the next
page to scrape for links.

A Parser and LinkSource implementation for FoodNetwork is added.